+++ allennlp/modules/elmo.py
@@ -2,17 +2,23 @@ import json
 from typing import Union, List, Dict
 
 import torch
+from torch.autograd import Variable
 import numpy
+import h5py
+
+from overrides import overrides
 
 from allennlp.common.file_utils import cached_path
 from allennlp.common.checks import ConfigurationError
 from allennlp.common import Registrable, Params
-from allennlp.modules.token_embedders.elmo_token_embedder import _ElmoTokenRepresentation
-from allennlp.modules.token_embedders import TokenEmbedder
 from allennlp.modules.elmo_lstm import ElmoLstm
-from allennlp.modules import ScalarMix
-from allennlp.nn.util import remove_sentence_boundaries
+from allennlp.modules.highway import Highway
+from allennlp.modules.scalar_mix import ScalarMix
+from allennlp.nn.util import remove_sentence_boundaries, add_sentence_boundary_token_ids
 from allennlp.data import Vocabulary
+from allennlp.data.token_indexers.elmo_indexer import ELMoCharacterMapper
+
+# pylint: disable=attribute-defined-outside-init
 
 
 @Registrable.register('elmo')
@@ -46,11 +52,10 @@ class Elmo(torch.nn.Module, Registrable):
                  options_file: str,
                  weight_file: str,
                  num_elmo_layers: int,
-                 do_layer_norm: bool=False) -> None:
+                 do_layer_norm: bool = False) -> None:
         super(Elmo, self).__init__()
 
         self._elmo_lstm = _ElmoBiLm(options_file, weight_file)
-        #self.add_module('elmo_lstm', self._elmo_lstm)
 
         self._scalar_mixes = []
         for k in range(num_elmo_layers):
@@ -99,14 +104,19 @@ class Elmo(torch.nn.Module, Registrable):
         return cls(options_file, weight_file, num_elmo_layers, do_layer_norm)
 
 
-@TokenEmbedder.register("elmo_token_embedder")
-class ElmoTokenEmbedder(TokenEmbedder):
+class _ElmoTokenRepresentation(torch.nn.Module):
     """
-    Compute a single layer of ELMo representations.
+    Compute context sensitive token representation using pretrained biLM.
 
-    This class servers as a convenience when you only want to use one layer of
-    ELMo representations at the input of your network.  It's essentially a wrapper
-    around Elmo(num_elmo_layers=1, ...)
+    This embedder has input character ids of size (batch_size, sequence_length, 50)
+    and returns (batch_size, sequence_length + 2, embedding_dim), where embedding_dim
+    is specified in the options file (typically 512).
+
+    We add special entries at the beginning and end of each sequence corresponding
+    to <S> and </S>, the beginning and end of sentence tokens.
+
+    Note: this is a lower level class useful for advanced usage.  Most users should
+    use ``ElmoTokenEmbedder`` or ``allennlp.modules.Elmo`` instead.
 
     Parameters
     ----------
@@ -114,43 +124,223 @@ class ElmoTokenEmbedder(TokenEmbedder):
         ELMo JSON options file
     weight_file : str
         ELMo hdf5 weight file
-    do_layer_norm: bool
-        Should we apply layer normalization (passed to ``ScalarMix``)?
+
+    The relevant section of the options file is something like:
+    .. example-code::
+
+        .. code-block:: python
+
+            {'char_cnn': {
+                'activation': 'relu',
+                'embedding': {'dim': 4},
+                'filters': [[1, 4], [2, 8], [3, 16], [4, 32], [5, 64]],
+                'max_characters_per_token': 50,
+                'n_characters': 262,
+                'n_highway': 2
+                }
+            }
+
+    something
     """
     def __init__(self,
                  options_file: str,
-                 weight_file: str,
-                 do_layer_norm: bool=False) -> None:
-        super(ElmoTokenEmbedder, self).__init__()
+                 weight_file: str) -> None:
+        super(_ElmoTokenRepresentation, self).__init__()
 
-        self._elmo = Elmo(options_file, weight_file, 1, do_layer_norm=do_layer_norm)
-        #self.add_module('elmo', self._elmo)
+        with open(cached_path(options_file), 'r') as fin:
+            self._options = json.load(fin)
+        self._weight_file = weight_file
+
+        self.output_dim = self._options['lstm']['projection_dim']
+
+        self._load_weights()
+
+        # Cache the arrays for use in forward -- +1 due to masking.
+        self._beginning_of_sentence_characters = Variable(torch.from_numpy(
+                numpy.array(ELMoCharacterMapper.beginning_of_sentence_characters) + 1
+        ))
+        self._end_of_sentence_characters = Variable(torch.from_numpy(
+                numpy.array(ELMoCharacterMapper.end_of_sentence_characters) + 1
+        ))
 
     def get_output_dim(self):
-        return 2 * self._elmo._elmo_lstm._token_embedder.get_output_dim()
+        return self.output_dim
 
-    def forward(self, inputs: torch.Tensor) -> torch.Tensor:
+    @overrides
+    def forward(self, inputs: torch.Tensor) -> Dict[str, torch.Tensor]:  # pylint: disable=arguments-differ
         """
+        Compute context insensitive token embeddings for ELMo representations.
+
         Parameters
         ----------
         inputs: ``torch.autograd.Variable``
-            Shape ``(batch_size, timesteps, 50)`` of character ids representing the current batch.
+            Shape ``(batch_size, sequence_length, 50)`` of character ids representing the
+            current batch.
 
         Returns
         -------
-        The ELMo representations for the input sequence, shape
-        ``(batch_size, timesteps, embedding_dim)``
+        Dict with keys:
+
+        ``'token_embedding'``: ``torch.autograd.Variable``
+            Shape ``(batch_size, sequence_length + 2, embedding_dim)`` tensor with context
+            insensitive token representations.
+        ``'mask'``:  ``torch.autograd.Variable``
+            Shape ``(batch_size, sequence_length + 2)`` long tensor with sequence mask.
         """
-        elmo_output = self._elmo(inputs)
-        return elmo_output['elmo'][0]
+        # Add BOS/EOS
+        mask = ((inputs > 0).long().sum(dim=-1) > 0).long()
+        character_ids_with_bos_eos, mask_with_bos_eos = add_sentence_boundary_token_ids(
+                inputs,
+                mask,
+                self._beginning_of_sentence_characters,
+                self._end_of_sentence_characters
+        )
+
+        # the character id embedding
+        max_chars_per_token = self._options['char_cnn']['max_characters_per_token']
+        # (batch_size * sequence_length, max_chars_per_token, embed_dim)
+        character_embedding = torch.nn.functional.embedding(
+                character_ids_with_bos_eos.view(-1, max_chars_per_token),
+                self._char_embedding_weights
+        )
+
+        # run convolutions
+        cnn_options = self._options['char_cnn']
+        if cnn_options['activation'] == 'tanh':
+            activation = torch.nn.functional.tanh
+        elif cnn_options['activation'] == 'relu':
+            activation = torch.nn.functional.relu
+        else:
+            raise ConfigurationError("Unknown activation")
+
+        # (batch_size * sequence_length, embed_dim, max_chars_per_token)
+        character_embedding = torch.transpose(character_embedding, 1, 2)
+        convs = []
+        for conv in self._convolutions:
+            convolved = conv(character_embedding)
+            # (batch_size * sequence_length, n_filters for this width)
+            convolved, _ = torch.max(convolved, dim=-1)
+            convolved = activation(convolved)
+            convs.append(convolved)
+
+        # (batch_size * sequence_length, n_filters)
+        token_embedding = torch.cat(convs, dim=-1)
+
+        # apply the highway layers (batch_size * sequence_length, n_filters)
+        token_embedding = self._highways(token_embedding)
+
+        # final projection  (batch_size * sequence_length, embedding_dim)
+        token_embedding = self._projection(token_embedding)
+
+        # reshape to (batch_size, sequence_length, embedding_dim)
+        batch_size, sequence_length, _ = character_ids_with_bos_eos.size()
+
+        return {
+                'mask': mask_with_bos_eos,
+                'token_embedding': token_embedding.view(batch_size, sequence_length, -1)
+        }
+
+    def _load_weights(self):
+        self._load_char_embedding()
+        self._load_cnn_weights()
+        self._load_highway()
+        self._load_projection()
+
+    def _load_char_embedding(self):
+        with h5py.File(cached_path(self._weight_file), 'r') as fin:
+            char_embed_weights = fin['char_embed'][...]
+
+        weights = numpy.zeros(
+                (char_embed_weights.shape[0] + 1, char_embed_weights.shape[1]),
+                dtype='float32'
+        )
+        weights[1:, :] = char_embed_weights
+
+        self._char_embedding_weights = torch.nn.Parameter(
+                torch.FloatTensor(weights), requires_grad=False
+        )
+
+    def _load_cnn_weights(self):
+        cnn_options = self._options['char_cnn']
+        filters = cnn_options['filters']
+        char_embed_dim = cnn_options['embedding']['dim']
+
+        convolutions = []
+        for i, (width, num) in enumerate(filters):
+            conv = torch.nn.Conv1d(
+                    in_channels=char_embed_dim,
+                    out_channels=num,
+                    kernel_size=width,
+                    bias=True
+            )
+            # load the weights
+            with h5py.File(cached_path(self._weight_file), 'r') as fin:
+                weight = fin['CNN']['W_cnn_{}'.format(i)][...]
+                bias = fin['CNN']['b_cnn_{}'.format(i)][...]
+
+            w_reshaped = numpy.transpose(weight.squeeze(axis=0), axes=(2, 1, 0))
+            if w_reshaped.shape != tuple(conv.weight.data.shape):
+                raise ValueError("Invalid weight file")
+            conv.weight.data.copy_(torch.FloatTensor(w_reshaped))
+            conv.bias.data.copy_(torch.FloatTensor(bias))
+
+            conv.weight.requires_grad = False
+            conv.bias.requires_grad = False
+
+            convolutions.append(conv)
+
+        self._convolutions = convolutions
+
+    def _load_highway(self):
+        # pylint: disable=protected-access
+        # the highway layers have same dimensionality as the number of cnn filters
+        cnn_options = self._options['char_cnn']
+        filters = cnn_options['filters']
+        n_filters = sum(f[1] for f in filters)
+        n_highway = cnn_options['n_highway']
+
+        # create the layers, and load the weights
+        self._highways = Highway(n_filters, n_highway, activation=torch.nn.functional.relu)
+        for k in range(n_highway):
+            # The AllenNLP highway is one matrix multplication with concatenation of
+            # transform and carry weights.
+            with h5py.File(cached_path(self._weight_file), 'r') as fin:
+                # The weights are transposed due to multiplication order assumptions in tf
+                # vs pytorch (tf.matmul(X, W) vs pytorch.matmul(W, X))
+                w_transform = numpy.transpose(fin['CNN_high_{}'.format(k)]['W_transform'][...])
+                # -1.0 since AllenNLP is g * x + (1 - g) * f(x) but tf is (1 - g) * x + g * f(x)
+                w_carry = -1.0 * numpy.transpose(fin['CNN_high_{}'.format(k)]['W_carry'][...])
+                weight = numpy.concatenate([w_transform, w_carry], axis=0)
+                self._highways._layers[k].weight.data.copy_(torch.FloatTensor(weight))
+                self._highways._layers[k].weight.requires_grad = False
+
+                b_transform = fin['CNN_high_{}'.format(k)]['b_transform'][...]
+                b_carry = -1.0 * fin['CNN_high_{}'.format(k)]['b_carry'][...]
+                bias = numpy.concatenate([b_transform, b_carry], axis=0)
+                self._highways._layers[k].bias.data.copy_(torch.FloatTensor(bias))
+                self._highways._layers[k].bias.requires_grad = False
+
+    def _load_projection(self):
+        cnn_options = self._options['char_cnn']
+        filters = cnn_options['filters']
+        n_filters = sum(f[1] for f in filters)
+
+        self._projection = torch.nn.Linear(n_filters, self.output_dim, bias=True)
+        with h5py.File(cached_path(self._weight_file), 'r') as fin:
+            weight = fin['CNN_proj']['W_proj'][...]
+            bias = fin['CNN_proj']['b_proj'][...]
+            self._projection.weight.data.copy_(torch.FloatTensor(numpy.transpose(weight)))
+            self._projection.bias.data.copy_(torch.FloatTensor(bias))
+
+            self._projection.weight.requires_grad = False
+            self._projection.bias.requires_grad = False
 
     @classmethod
-    def from_params(cls, vocab: Vocabulary, params: Params) -> 'ElmoTokenEmbedder':
+    def from_params(cls, vocab: Vocabulary, params: Params) -> '_ElmoTokenRepresentation':
         options_file = params.pop('options_file')
         weight_file = params.pop('weight_file')
-        do_layer_norm = params.pop('do_layer_norm', False)
         params.assert_empty(cls.__name__)
-        return cls(options_file, weight_file, do_layer_norm)
+        return cls(options_file, weight_file)
 
 
 class _ElmoBiLm(torch.nn.Module):
@@ -174,20 +364,18 @@ class _ElmoBiLm(torch.nn.Module):
         super(_ElmoBiLm, self).__init__()
 
         self._token_embedder = _ElmoTokenRepresentation(options_file, weight_file)
-        #self.add_module('elmo_token_embedder', self._token_embedder)
 
         with open(cached_path(options_file), 'r') as fin:
             options = json.load(fin)
         if not options['lstm'].get('use_skip_connections'):
             raise ConfigurationError('We only support pretrained biLMs with residual connections')
         self._elmo_lstm = ElmoLstm(input_size=options['lstm']['projection_dim'],
-                                hidden_size=options['lstm']['projection_dim'],
-                                cell_size=options['lstm']['dim'],
-                                num_layers=options['lstm']['n_layers'],
-                                memory_cell_clip_value=options['lstm']['cell_clip'],
-                                state_projection_clip_value=options['lstm']['proj_clip'])
+                                   hidden_size=options['lstm']['projection_dim'],
+                                   cell_size=options['lstm']['dim'],
+                                   num_layers=options['lstm']['n_layers'],
+                                   memory_cell_clip_value=options['lstm']['cell_clip'],
+                                   state_projection_clip_value=options['lstm']['proj_clip'])
         self._elmo_lstm._load_weights(weight_file)
-        #self.add_module('elmo_lstm', self._elmo_lstm)
         # Number of representation layers including context independent layer
         self.num_layers = options['lstm']['n_layers'] + 1
