+++ allennlp/modules/token_embedders/__init__.py
@@ -6,3 +6,4 @@ embeds one-hot-encoded tokens as vectors.
 from allennlp.modules.token_embedders.token_embedder import TokenEmbedder
 from allennlp.modules.token_embedders.embedding import Embedding
 from allennlp.modules.token_embedders.token_characters_encoder import TokenCharactersEncoder
+from allennlp.modules.token_embedders.elmo_token_embedder import ElmoTokenEmbedder
