+++ tests/modules/elmo_test.py
@@ -14,7 +14,7 @@ from allennlp.nn.util import arrays_to_variables
 from allennlp.data.token_indexers.elmo_indexer import ELMoTokenCharactersIndexer
 from allennlp.data import Token, Vocabulary, Dataset, Instance
 from allennlp.data.iterators import BasicIterator
-from allennlp.modules.elmo import _ElmoBiLm, Elmo
+from allennlp.modules.elmo import _ElmoBiLm, Elmo, _ElmoTokenRepresentation
 from allennlp.data.fields import TextField
 
 FIXTURES = os.path.join('tests', 'fixtures', 'elmo')
@@ -132,23 +132,52 @@ class TestElmo(AllenNlpTestCase):
         assert list(mask.size()) == [2, 7]
 
 
-class TestElmoTokenEmbedder(ModelTestCase):
-    def setUp(self):
-        super().setUp()
-        self.set_up_model('tests/fixtures/elmo/config/characters_token_embedder.json',
-                          'tests/fixtures/data/conll2003.txt')
+class TestElmoTokenRepresentation(AllenNlpTestCase):
+    def test_elmo_token_representation(self):
+        # Load the test words and convert to char ids
+        with open(os.path.join(FIXTURES, 'vocab_test.txt'), 'r') as fin:
+            tokens = fin.read().strip().split('\n')
 
-    def test_tagger_with_elmo_token_embedder_can_train_save_and_load(self):
-        self.ensure_model_can_train_save_and_load(self.param_file)
+        indexer = ELMoTokenCharactersIndexer()
+        indices = [indexer.token_to_indices(Token(token), Vocabulary()) for token in tokens]
+        # There are 457 tokens. Reshape into 10 batches of 50 tokens.
+        sentences = []
+        for k in range(10):
+            sentences.append(
+                    indexer.pad_token_sequence(
+                            indices[(k * 50):((k + 1) * 50)], desired_num_tokens=50, padding_lengths={}
+                    )
+            )
+        batch = Variable(torch.from_numpy(numpy.array(sentences)))
 
-    def test_tagger_with_elmo_token_embedder_forward_pass_runs_correctly(self):
-        training_arrays = self.dataset.as_array_dict()
-        output_dict = self.model.forward(**arrays_to_variables(training_arrays))
-        tags = output_dict['tags']
-        assert len(tags) == 2
-        assert len(tags[0]) == 7
-        assert len(tags[1]) == 7
-        for example_tags in tags:
-            for tag_id in example_tags:
-                tag = self.model.vocab.get_token_from_index(tag_id, namespace="labels")
-                assert tag in {'O', 'I-ORG', 'I-PER', 'I-LOC'}
+        options_file = os.path.join(FIXTURES, 'options.json')
+        weight_file = os.path.join(FIXTURES, 'lm_weights.hdf5')
+
+        elmo_token_embedder = _ElmoTokenRepresentation(options_file, weight_file)
+        token_embedding = elmo_token_embedder(batch)['token_embedding'].data.numpy()
+
+        # Reshape back to a list of words and compare with ground truth.  Need to also
+        # remove <S>, </S>
+        actual_embeddings = token_embedding[:, 1:-1, :].reshape(-1, token_embedding.shape[-1])
+
+        embedding_file = os.path.join(FIXTURES, 'elmo_token_embeddings.hdf5')
+        with h5py.File(embedding_file, 'r') as fin:
+            expected_embeddings = fin['embedding'][...]
+
+        assert numpy.allclose(actual_embeddings[:len(tokens)], expected_embeddings, atol=1e-6)
+
+    def test_elmo_token_representation_bos_eos(self):
+        # The additional <S> and </S> embeddings added by the embedder should be as expected.
+        indexer = ELMoTokenCharactersIndexer()
+
+        options_file = os.path.join(FIXTURES, 'options.json')
+        weight_file = os.path.join(FIXTURES, 'lm_weights.hdf5')
+
+        elmo_token_embedder = _ElmoTokenRepresentation(options_file, weight_file)
+
+        # First <S>
+        for correct_index, token in [[0, '<S>'], [2, '</S>']]:
+            indices = indexer.token_to_indices(Token(token), Vocabulary())
+            indices = Variable(torch.from_numpy(numpy.array(indices))).view(1, 1, -1)
+            embeddings = elmo_token_embedder(indices)['token_embedding']
+            assert numpy.allclose(embeddings[0, correct_index, :].data.numpy(), embeddings[0, 1, :].data.numpy())
